feat: move menu components into componets folder and then better structured the menu items
