+++ src/ui/components/common/header/notification-bell/NotificationBell.css
