+++ src/ui/config/menu/settingsRootMenuItem.ts
@@ -1,0 +1,16 @@
+import type { Priority, RootMenuItem } from "../../providers/MenuProvider";
+
+export function createSettingsRootMenuItem(
+    action: () => void,
+    priority: Priority = 9,
+): RootMenuItem {
+    return {
+        id: "settings",
+        title: "Settings",
+        task: {
+            action: action,
+            type: "direct",
+        },
+        priority: priority,
+    };
+}
