+++ src/ui/layouts/MainLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import { Footer } from "../components/common/footer/Footer";
-import { Header } from "./header/Header";
+import { Header } from "../components/common/header/Header";
 import { MainMenuBridge } from "./menu-bridges/MainMenuBridge";
 import { ServerStatus } from "../features/system/components/server-status/ServerStatus";
