+++ src/ui/main.tsx
@@ -1,14 +1,14 @@
 import { RouterProvider } from "react-router-dom";
 import { createRoot } from "react-dom/client";
+import { router } from "./router/router.ts";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { UserSettingsProvider } from "./providers/UserSettingsProvider.tsx";
 import { AppearanceProvider } from "./providers/AppearanceProvider.tsx";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { router } from "./router/router.ts";
+import { DialogueHost } from "./components/common/dialogue/DialogueHost.tsx";
 
 import "@mantine/core/styles.css";
 import "@mantine/notifications/styles.css";
 import "./index.css";
-import { DialogueHost } from "./components/common/dialogue/DialogueHost.tsx";
 
 const queryClient = new QueryClient();
